pyproject.toml
[project]
name = "the-basic-support-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.136.1",
    "langchain>=1.2.17",
    "langchain-openai>=1.2.1",
    "langgraph>=1.1.10",
    "pydantic-settings>=2.14.0",
    "uvicorn>=0.46.0",
]
